Update the header and split up components
