+++ client-course-schedulizer/src/components/Header/Header/Header.scss
@@ -1,19 +1,32 @@
 .toolbar {
-  display: flex;
   justify-content: space-between;
 }
 
 .org-logo {
   width: 60px;
+  height: auto;
 }
 
-.hidden {
-  display: none !important;
+.toolbar-group {
+  align-items: center;
+  display: flex;
+  flex-direction: row;
 }
 
-.import-menu-item,
-.export-menu-item {
-  padding: 0px !important;
+// Style all children of .links parent
+.links > * {
+  margin-left: 1em !important;
+  color: white !important;
+}
+
+// Style all children of .title parent
+.title > * {
+  margin-right: 1em !important;
+}
+
+.title {
+  text-decoration: none;
+  color: white;
 }
 
 .import-label,
