+++ client-course-schedulizer/src/components/Header/Header/Header.tsx
@@ -1,45 +1,31 @@
-import { AppBar, IconButton, Menu, MenuItem, Toolbar, Typography } from "@material-ui/core";
-import { Menu as MenuIcon } from "@material-ui/icons";
-import { bindMenu, bindTrigger, usePopupState } from "material-ui-popup-state/hooks";
+import { AppBar, Toolbar, Typography } from "@material-ui/core";
 import React from "react";
-import logo from "../../../assets/CalvinUniv-vert-full-color-inverse.png";
-import { ImportButton } from "../ImportButton";
-import { ExportButton } from "../ExportButton";
+import { Link as RouterLink } from "react-router-dom";
+import Link from "@material-ui/core/Link";
+import { logo } from "../../../assets";
 import "./Header.scss";
 
+/* Header is the top bar that persists across the app */
 export const Header = () => {
-  const popupState = usePopupState({ popupId: "menu", variant: "popover" });
-
   return (
-    <div className="header">
-      <AppBar position="static">
-        <Toolbar className="toolbar">
-          <IconButton color="inherit" edge="start" {...bindTrigger(popupState)}>
-            <MenuIcon />
-          </IconButton>
-          <Menu // Anchoring from: https://codesandbox.io/s/3rmgv?file=/demo.js:603-812
-            anchorOrigin={{
-              horizontal: "right",
-              vertical: "top",
-            }}
-            getContentAnchorEl={null}
-            transformOrigin={{
-              horizontal: "left",
-              vertical: "top",
-            }}
-            {...bindMenu(popupState)}
-          >
-            <MenuItem className="import-menu-item">
-              <ImportButton />
-            </MenuItem>
-            <MenuItem className="export-menu-item">
-              <ExportButton />
-            </MenuItem>
-          </Menu>
-          <Typography variant="h6">Course Schedulizer</Typography>
+    <AppBar position="static">
+      <Toolbar className="toolbar">
+        <RouterLink className="toolbar-group title" to="/">
           <img alt="Org Logo" className="org-logo" src={logo} />
-        </Toolbar>
-      </AppBar>
-    </div>
+          <Typography variant="h6">Course Schedulizer</Typography>
+        </RouterLink>
+        <Typography className="toolbar-group links">
+          <Link component={RouterLink} to="/">
+            Schedulizer
+          </Link>
+          <Link component={RouterLink} to="/about">
+            About
+          </Link>
+          <Link component={RouterLink} to="/help">
+            Help
+          </Link>
+        </Typography>
+      </Toolbar>
+    </AppBar>
   );
 };
