+++ client-course-schedulizer/src/components/Header/ImportButton/ImportButton.scss
@@ -1,0 +1,3 @@
+.hidden {
+  display: none !important;
+}
