+++ client-course-schedulizer/src/components/Header/ImportButton/ImportButton.tsx
@@ -1,6 +1,7 @@
 import { Input, InputLabel } from "@material-ui/core";
 import React from "react";
 import { useImportFile } from "../../../utilities/hooks/useImportFile";
+import "./ImportButton.scss";
 
 interface ImportButton {
   className?: string;
