+++ client-course-schedulizer/src/components/Tabs/tabComponents/CSVActions/CSVActions.scss
@@ -1,0 +1,4 @@
+.import-menu-item,
+.export-menu-item {
+  padding: 0px !important;
+}
