+++ client-course-schedulizer/src/components/Tabs/tabComponents/NoCoursesHeader/NoCoursesHeader.scss
@@ -1,0 +1,4 @@
+.import-button {
+  margin-right: 1em !important;
+  color: white !important;
+}
