+++ client-course-schedulizer/src/components/Tabs/tabComponents/NoCoursesHeader/index.ts
@@ -1,0 +1,1 @@
+export * from "./NoCoursesHeader";
