+++ client-course-schedulizer/src/components/Tabs/tabComponents/TabPanel/index.ts
@@ -1,0 +1,1 @@
+export * from "./TabPanel";
