+++ client-course-schedulizer/src/components/Tabs/tabComponents/index.ts
@@ -1,0 +1,3 @@
+export * from "./TabPanel";
+export * from "./NoCoursesHeader";
+export * from "./CSVActions";
